created class 'Clan'; created test envs for vscode and jupyter; changed 'beacon' to 'flare'; updated gitignore to include __pycache__/
