+++ .gitignore
@@ -1,3 +1,4 @@
+config.ini
+__pycache__/
 venv/
 .vscode/
-config.ini
